requirements.txt
# Web Framework
Flask==3.0.0
Werkzeug==3.0.1
Flask-WTF==1.2.1

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
pyodbc==5.0.1
sqlalchemy-pytds==1.0.2

# Authentication
Flask-Login==0.6.3

# API and Data Processing
requests==2.31.0
python-dateutil==2.8.2
pandas==2.1.4
numpy==1.26.2

# Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Security
Flask-Talisman==1.1.0
Flask-SeaSurf==1.1.1
